+++ test/test_league_loader/test_YahooLeagueLoader.py
@@ -15,6 +15,13 @@ from leeger.model.league.Year import Year
 
 class TestYahooLeagueLoader(unittest.TestCase):
     # helper methods
+    
+    def __foo(self) -> callable:
+        def mockTeamsMethod():
+            return ["", "123"]
+
+        return mockTeamsMethod
+    
     def __getMockTeamsMethod(self, teams: list) -> callable:
         def mockTeamsMethod():
             return teams
@@ -117,7 +124,7 @@ class TestYahooLeagueLoader(unittest.TestCase):
         mockLeague2023.season = 2023
         mockLeague2023.current_week = 3
         mockLeague2023.end_week = 3
-        mockLeague2023.past_league_id = "123"
+        mockLeague2023.past_league_id = ["", "123"]
 
         # mock fake leagues 2023
         mockLeague2023_fake1 = Mock()
@@ -665,7 +672,7 @@ class TestYahooLeagueLoader(unittest.TestCase):
         mockLeague2023.season = 2023
         mockLeague2023.current_week = 3
         mockLeague2023.end_week = 3
-        mockLeague2023.past_league_id = "123"
+        mockLeague2023.past_league_id = ["", "123"]
 
         # mock fake leagues 2023
         mockLeague2023_fake1 = Mock()
@@ -1226,7 +1233,7 @@ class TestYahooLeagueLoader(unittest.TestCase):
         mockLeague2023.season = 2023
         mockLeague2023.current_week = 3
         mockLeague2023.end_week = 3
-        mockLeague2023.past_league_id = "123"
+        mockLeague2023.past_league_id = ["", "123"]
 
         # mock fake leagues 2023
         mockLeague2023_fake1 = Mock()